+++ NEWS.md
@@ -8,9 +8,17 @@
 
 * `if.()`, `else_if.()`: Are now also able to use the new writing style with conditions as characters introduced by `ifelse_multi()`. (20.06.2026)
 
+### Fixed
+
+* `multi_join()`: Before joining it is now checked whether there will be duplicate variable names after the join. If so, these variables will be dropped before joining. Otherwise it was possible that a variable could show up in the final data frame with the exact same name multiple times. (21.06.2026)
+
 ### Optimization
 
 * `summarise_plus()`: Brought percentiles up to more speed. (17.06.2026)
+
+### Additionally
+
+* `multi_join()`: Now displays the actual data frame names to be joined instead of just an iterative number in the console. (21.06.2026)
 
 # qol 1.3.2